refactor: move press system files to air conditioning
