+++ fbw-common/src/wasm/systems/systems/src/lib.rs
@@ -17,6 +17,5 @@ pub mod landing_gear;
 pub mod navigation;
 pub mod overhead;
 pub mod pneumatic;
-pub mod pressurization;
 pub mod shared;
 pub mod simulation;
